Move MyFirstClass to util package.
