+++ app/src/main/java/com/example/myfirstapp/MainActivity.kt
@@ -2,18 +2,24 @@ package com.example.myfirstapp
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.example.myfirstapp.utilities.MyClass
+import com.example.myfirstapp.util.MyFirstClass
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        val myClass = MyClass()
 
-        myClass.saySomething()
+        saySomething()
         // Command Backspace to delete
         // Command D to duplicate
 
+    }
+
+    private fun saySomething() {
+        val myFirstClass = MyFirstClass()
+
+
+        myFirstClass.saySomething()
     }
 
     fun myFunc1() {
