+++ app/src/main/java/com/example/myfirstapp/util/MyFirstClass.kt
@@ -1,8 +1,9 @@
-package com.example.myfirstapp.utilities
+package com.example.myfirstapp.util
 
 import android.util.Log
 const val TAG = "my_tag"
-class MyClass {
+class MyFirstClass {
+    // Shift F6 renames the variable
     fun saySomething() {
         Log.d(TAG, "saySomething: Hello!")
     }
